Small functionality adjustments/additions
ElementPanel can now remove images from its cache so they can be updated
DrawnAnimation uses a double for scale now for better tuned display
CanvasPanel can now adjustSize and export the image it is displaying at regular or scaled sizes
WindowFrame had oversight for adding a panel to a non-existent Window which forgot to set that window to the active one if no others existed, fixed by piggybacking the existing reserveWindow function.

+++ visual/panel/element/DrawnAnimation.java
@@ -17,9 +17,9 @@ public class DrawnAnimation extends Element{
 	
 	private Image[] images;
 	
-	private int scale;
+	private double scale;
 
-	public DrawnAnimation(int x, int y, int prior, int wait, boolean center, int inScale, Image[] imgs) {
+	public DrawnAnimation(int x, int y, int prior, int wait, boolean center, double inScale, Image[] imgs) {
 		setX(x);
 		setY(y);
 		setDrawPriority(prior);
@@ -34,7 +34,7 @@ public class DrawnAnimation extends Element{
 		scaleImages();
 	}
 	
-	public DrawnAnimation(int x, int y, int prior, int[] wait, boolean center, int inScale, Image[] imgs) {
+	public DrawnAnimation(int x, int y, int prior, int[] wait, boolean center, double inScale, Image[] imgs) {
 		setX(x);
 		setY(y);
 		setDrawPriority(prior);
